Implemented mmap functions

+++ src/io/lseek.c
@@ -6,7 +6,7 @@
 
 off_t lseek(int fd, off_t offset, int whence)
 {
-	int call = syscall(__NR_lseek, fd, offset, whence);
+	off_t call = syscall(__NR_lseek, fd, offset, whence);
   if (call < 0) {
     errno = -call;
     return -1;
